LcTerminalPdb extends TerminalPdb to handling list comprehensions using local variables in closure
LcTerminalPdb updates the default() method from TerminalPdb (which uses Pdb's
method). default() is called when code is run from the prompt, and uses exec()
to evaluate the code with the frame's local and global variables. When this
code is a list comprehension (or other closure) that requires local variables
to be bound, exec() fails with a NameError. This is because closures using
exec() can apparently only access globals. Here, when a NameError occurs,
LcTerminalPdb.default attempts to rerun the code, but by first updating the
global dictionary passed to exec to include all the local variables. The
dictionary is first copied to avoid any weird updates to frame globals, and the
code is written such that processing of any other commands happens as normal.